+++ src/Telegram.Bot/Types/ReplyMarkups/KeyboardButton.cs
@@ -1,7 +1,7 @@
 ﻿using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 
-namespace Telegram.Bot.Types.ReplyMarkups.Buttons
+namespace Telegram.Bot.Types.ReplyMarkups
 {
     /// <summary>
     /// This object represents one button of the reply keyboard. For simple text buttons String can be used instead of this object to specify text of the button.
@@ -11,7 +11,26 @@ namespace Telegram.Bot.Types.ReplyMarkups.Buttons
     {
         /// <inheritdoc />
         [JsonProperty(Required = Required.Always)]
-        public string Text { get; }
+        public string Text { get; set; }
+
+        /// <summary>
+        /// If <c>true</c>, the user's current location will be sent when the button is pressed. Available in private chats only
+        /// </summary>
+        [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
+        public bool RequestLocation { get; set; }
+
+        /// <summary>
+        /// If <c>true</c>, the user's phone number will be sent as a contact when the button is pressed. Available in private chats only
+        /// </summary>
+        [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
+        public bool RequestContact { get; set; }
+
+        /// <summary>
+        /// Initializes a new <see cref="KeyboardButton"/>
+        /// </summary>
+        public KeyboardButton()
+        {
+        }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="KeyboardButton"/> class.
@@ -21,6 +40,12 @@ namespace Telegram.Bot.Types.ReplyMarkups.Buttons
         {
             Text = text;
         }
+
+        public static KeyboardButton WithRequestContact(string text) =>
+            new KeyboardButton(text) {RequestContact = true};
+
+        public static KeyboardButton WithRequestLocation(string text) =>
+            new KeyboardButton(text) {RequestLocation = true};
 
         /// <summary>
         /// Performs an implicit conversion from <see cref="string"/> to <see cref="KeyboardButton"/>.